formatting and minor changes in create shelf functionality

+++ src/app/services/shelf-service.ts
@@ -11,9 +11,9 @@ export class ShelfService {
   constructor(private httpClient:HttpClient) {}
   baseURL:string = "http://localhost:8080"
 
-  currentShelfPosition:ShelfPosition = new ShelfPosition()
+  currentShelfPosition:ShelfPosition|null = null
 
-  getCurrentShelfPosition():ShelfPosition {
+  getCurrentShelfPosition():ShelfPosition|null {
     return this.currentShelfPosition
   }
 
@@ -21,11 +21,13 @@ export class ShelfService {
     this.currentShelfPosition = shelfPosition;
   }
 
-  createShelf(shelf:Shelf):Observable<Shelf> {
+  createShelf(shelf:Shelf):Observable<Shelf>|null {
+    if(this.currentShelfPosition == null) return null
     return this.httpClient.post<Shelf>(`${this.baseURL}/shelf/create?shelfPositionId=${this.currentShelfPosition.id}`, shelf)
   }
 
-  deleteShelf(shelf:Shelf) {
+  deleteShelf(shelf:Shelf):void|null {
+    if(this.currentShelfPosition == null) return null
     this.httpClient.delete(`${this.baseURL}/shelf/delete?shelfId=${shelf.id}`)
     console.log("shelf with id: " + shelf.id + " deleted")
   }
